requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
compressed-tensors-nightly==0.9.3.20250408
datasets==4.0.0
dill==0.3.8
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.9
huggingface-hub==0.35.0rc0
idna==3.10
Jinja2==3.1.6
llmcompressor-nightly==0.5.0.20250410
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-modelopt==0.33.1
nvidia-modelopt-core==0.33.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.9.79
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
PuLP==3.2.2
pyarrow==21.0.0
pydantic==2.12.0a1
pydantic_core==2.37.2
Pygments==2.19.2
pynvml==12.0.0
python-dateutil==2.9.0.post0
pytorch-triton==3.4.0+gitf7888497
pytz==2025.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scipy==1.16.1
setuptools==78.1.0
six==1.17.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.9.0.dev20250901+cu129
torchprofile==0.0.4
torchvision==0.24.0.dev20250901+cu129
tqdm==4.67.1
transformers==4.49.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
